make it compile warning free
svn path=/trunk/KDE/kdelibs/; revision=623409

+++ khtml/misc/helper.cpp
@@ -70,7 +70,10 @@ QRgb khtml::qRgbaFromHsla(double h, double s, double l, double a)
     double temp2 = l < 0.5 ? l * (1.0 + s) : l + s - l * s;
     double temp1 = 2.0 * l - temp2;
 
-    return qRgba(calcHue(temp1, temp2, h + 1.0 / 3.0) * 255, calcHue(temp1, temp2, h) * 255, calcHue(temp1, temp2, h - 1.0 / 3.0) * 255, a * 255);
+    return qRgba(static_cast<int>(calcHue(temp1, temp2, h + 1.0 / 3.0) * 255),
+                 static_cast<int>(calcHue(temp1, temp2, h) * 255),
+                 static_cast<int>(calcHue(temp1, temp2, h - 1.0 / 3.0) * 255),
+                 static_cast<int>(a * 255));
 }
 
 /** finds out the background color of an element
